app/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
psycopg2-binary==2.9.9
redis==5.0.4
pydantic==2.7.1
prometheus-client==0.20.0
httpx==0.27.0
pytest==8.2.0
pytest-asyncio==0.23.7
